change PersonEntity.id from int to long

+++ backend/src/main/kotlin/com/juko/itmo/infsys/data/repository/PersonRepository.kt
@@ -3,4 +3,4 @@ package com.juko.itmo.infsys.data.repository
 import com.juko.itmo.infsys.data.entity.PersonEntity
 import org.springframework.data.jpa.repository.JpaRepository
 
-interface PersonRepository : JpaRepository<PersonEntity, Int>
+interface PersonRepository : JpaRepository<PersonEntity, Long>
